# docs/pro/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Game Panel PRO API",
    "version": "2.0.59",
    "description": "Scoped API. UUIDs remain canonical; server paths also accept global numbers. Current owner privileges are rechecked per request. Native provisioning creates stopped services with central IDs, published templates and per-token budgets. Unknown outcomes are reconciled without automatic re-dispatch. See API-V1-GUIDE.md."
  },
  "security": [
    {
      "apiToken": []
    }
  ],
  "paths": {
...
    "/api/v1/servers/{id}/game-admins": {
      "get": {
        "operationId": "listGameAdmins",
        "summary": "Read AMXX Steam ID admins",
        "description": "Also requires fs.read. CS1.6/ReHLDS with AMXX installed. ETag identifies version. Custom/password entries are counted as unmanaged, never exposed.",
        "x-scopes": [
          "game-admins.read"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/GameAdmins"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/servers/{id}/game-admins/{steamId}": {
      "put": {
        "operationId": "upsertGameAdmin",
        "summary": "Change Steam ID admin",
        "description": "Also fs.read/fs.write. Quoted ETag required in If-Match. Atomic versioned write with recovery snapshot. Preserves comments/custom entries, rejects ambiguous authentication. Missing version 428; conflict 409. No console reload is issued.",
        "x-scopes": [
          "game-admins.write"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/GameAdmins"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          },
          "428": {
            "$ref": "#/components/responses/Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "flags": {
                    "type": "string",
                    "pattern": "^[a-y]{1,25}$"
                  }
                },
                "required": [
                  "flags"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "steamId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^STEAM_[0-5]:[01]:[0-9]{1,12}$"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "delete": {
        "operationId": "removeGameAdmin",
        "summary": "Change Steam ID admin",
        "description": "Also fs.read/fs.write. Quoted ETag required in If-Match. Atomic versioned write with recovery snapshot. Preserves comments/custom entries, rejects ambiguous authentication. Missing version 428; conflict 409. No console reload is issued.",
        "x-scopes": [
          "game-admins.write"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/GameAdmins"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          },
          "428": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "steamId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^STEAM_[0-5]:[01]:[0-9]{1,12}$"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "apiToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "gpp_ token"
      }
    },
    "schemas": {
...
      "GameAdmins": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "provider": {
            "const": "amxx"
          },
          "entries": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": false,
              "properties": {
                "steamId": {
                  "type": "string"
                },
                "flags": {
                  "type": "string"
                }
              }
            }
          },
          "unmanaged": {
            "type": "integer"
          },
          "version": {
            "type": "string"
          },
          "snapshotId": {
            "type": "string"
          },
          "warning": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "Error": {
        "description": "Structured error; requestId also appears in X-Request-ID",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "error",
                "requestId"
              ],
              "properties": {
                "error": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "requestId": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      }
    }
  }
}